# Ensure that each time series has one and only one metric id.