--- a/app/vmui/Dockerfile-web
+++ b/app/vmui/Dockerfile-web
@@ -1,4 +1,4 @@
-FROM golang:1.22.3 as build-web-stage
+FROM golang:1.22.4 as build-web-stage
 COPY build /build
 
 WORKDIR /build