--- a/app/vmui/packages/vmui/src/components/Main/Modal/Modal.tsx
+++ b/app/vmui/packages/vmui/src/components/Main/Modal/Modal.tsx
@@ -88,9 +88,11 @@ const Modal: FC<ModalProps> = ({
             </Button>
           </div>
         </div>
+        {/* tabIndex to fix Ctrl-A */}
         <div
           className="vm-modal-content-body"
           onMouseDown={handleMouseDown}
+          tabIndex={0}
         >
           {children}
         </div>