--- a/app/vmui/packages/vmui/src/setupProxy.js
+++ b/app/vmui/packages/vmui/src/setupProxy.js
@@ -1,0 +1,11 @@
+const { createProxyMiddleware } = require('http-proxy-middleware');
+
+module.exports = function (app) {
+  app.use(
+    "/api",
+    createProxyMiddleware({
+      target: "http://localhost:8490/api",
+      changeOrigin: true,
+    })
+  );
+};