--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/VictoriaMetrics/VictoriaMetrics
 
-go 1.22.3
+go 1.22.4
 
 require (
 	cloud.google.com/go/storage v1.41.0