--- a/lib/fs/dir_remover.go
+++ b/lib/fs/dir_remover.go
@@ -74,9 +74,12 @@ func isStaleNFSFileHandleError(err error) bool {
 }
 
 func isTemporaryNFSError(err error) bool {
-	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/61 for details.
+	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/61
+	// and https://github.com/VictoriaMetrics/VictoriaMetrics/issues/6396 for details.
 	errStr := err.Error()
-	return strings.Contains(errStr, "directory not empty") || strings.Contains(errStr, "device or resource busy")
+	return strings.Contains(errStr, "directory not empty") ||
+		strings.Contains(errStr, "device or resource busy") ||
+		strings.Contains(errStr, "file exists")
 }
 
 // MustStopDirRemover must be called in the end of graceful shutdown