--- a/lib/logstorage/block_result.go
+++ b/lib/logstorage/block_result.go
@@ -3,6 +3,8 @@ package logstorage
 import (
 	"math"
 	"slices"
+	"strconv"
+	"strings"
 	"sync/atomic"
 	"time"
 	"unsafe"
@@ -1914,6 +1916,50 @@ func getCanonicalColumnName(columnName string) string {
 		return "_msg"
 	}
 	return columnName
+}
+
+func tryParseNumber(s string) (float64, bool) {
+	if len(s) == 0 {
+		return 0, false
+	}
+	f, ok := tryParseFloat64(s)
+	if ok {
+		return f, true
+	}
+	nsecs, ok := tryParseDuration(s)
+	if ok {
+		return float64(nsecs), true
+	}
+	bytes, ok := tryParseBytes(s)
+	if ok {
+		return float64(bytes), true
+	}
+	if isLikelyNumber(s) {
+		f, err := strconv.ParseFloat(s, 64)
+		if err == nil {
+			return f, true
+		}
+		n, err := strconv.ParseInt(s, 0, 64)
+		if err == nil {
+			return float64(n), true
+		}
+	}
+	return 0, false
+}
+
+func isLikelyNumber(s string) bool {
+	if !isNumberPrefix(s) {
+		return false
+	}
+	if strings.Count(s, ".") > 1 {
+		// This is likely IP address
+		return false
+	}
+	if strings.IndexByte(s, ':') >= 0 || strings.Count(s, "-") > 2 {
+		// This is likely a timestamp
+		return false
+	}
+	return true
 }
 
 var nan = math.NaN()