… app/vlselect/vmui/asset-manifest.json
… app/vlselect/vmui/index.html
… app/vlselect/vmui/static/css/main.1c2f4a93.css
… app/vlselect/vmui/static/css/main.7e9644cd.css
… app/vlselect/vmui/static/js/main.bfdd2931.js
… app/vlselect/vmui/static/js/main.bfdd2931.js.LICENSE.txt
… app/vlselect/vmui/static/js/main.c3285306.js
… app/vlselect/vmui/static/media/MetricsQL.cb83d071da309a358bc0.md
… app/vmselect/vmui/asset-manifest.json
… app/vmselect/vmui/index.html
… app/vmselect/vmui/static/css/main.d0b400e3.css
… app/vmselect/vmui/static/css/main.fce049bf.css
… app/vmselect/vmui/static/js/main.59c17910.js
… app/vmselect/vmui/static/js/main.ff9af61b.js
… app/vmselect/vmui/static/js/main.ff9af61b.js.LICENSE.txt
… app/vmselect/vmui/static/media/MetricsQL.cb83d071da309a358bc0.md
… app/vmui/packages/vmui/package.json
… app/vmui/packages/vmui/src/assets/MetricsQL.md
… app/vmui/packages/vmui/src/pages/ExploreLogs/ExploreLogsBody/GroupLogs.tsx
… app/vmui/packages/vmui/src/pages/ExploreLogs/ExploreLogsBody/style.scss
… deployment/docker/docker-compose-victorialogs.yml
… deployment/docker/victorialogs/filebeat-docker/docker-compose.yml
… deployment/docker/victorialogs/filebeat-syslog/docker-compose.yml
… deployment/docker/victorialogs/fluentbit-docker/docker-compose.yml
… deployment/docker/victorialogs/logstash/docker-compose.yml
… deployment/docker/victorialogs/promtail/docker-compose.yml
… deployment/docker/victorialogs/vector-docker/docker-compose.yml
… deployment/logs-benchmark/docker-compose-elk.yml
… deployment/logs-benchmark/docker-compose-loki.yml
… deployment/logs-benchmark/docker-compose.yml