### Describe Your Changes

This change adresses the case when more than one metric ID is created for a single time series. This happens when data for a new time series is inserted concurrently. An example use case is when a lot of historical data is imported into the database.

For example, say we have a year of temperature registered across the globe:

```
# day 1

T{la="90", lo="0"} -10 2023-01-01
T{la="90", lo="0.25"} -10 2023-01-01
...
T{la="-90", lo="359.75"} -10 2023-01-01

# day 2

T{la="90", lo="0"} -9 2023-01-02
T{la="90", lo="0.25"} -9 2023-01-02
...
T{la="-90", lo="359.75"} -9 2023-01-02

# day 365

T{la="90", lo="0"} -11 2023-12-31
T{la="90", lo="0.25"} -11 2023-12-31
...
T{la="-90", lo="359.75"} -11 2023-12-31
```

If we import each day concurrently, then the same new time series will be imported concurrently. Since these time series do not exist in db yet, they are not present in tsdbCache nor in indexdb, which means that the slow insertion path will take place for them (see Storage.add() func) and each metric will get a new metric ID even though the metric has already been seen.

This is an uncommon use case and is should not be observerd under normal circumstances.

The key precondition here is that new time series need to be inserted concurrently, if they are inserted sequentially or concurrently but within one day, duplicates won't be created. The `TestEachTimeSeriesHasOneUniqueMetricID` addresses these three cases.

To show the effect of such concurrent insertion, the patch also provides a benchmark. Here are the results before the fix (the number of duplicates grows as concurrency grows):

```
user@laptop:~/p/github.com/rtm0/VictoriaMetrics/01/src$ GOMAXPROCS=32 go test -run=^$ -bench=StorageAddHistoricalRowsConcurrently ./lib/storage/ -loggerLevel=ERROR; sync
goos: linux
goarch: amd64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/lib/storage
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkStorageAddHistoricalRowsConcurrently/1-32            30          40288038 ns/op           2482126 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/2-32            24          41880760 ns/op           2387731 rows/s             1766 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/4-32            30          38699310 ns/op           2584025 rows/s             1930 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/8-32            31          36784672 ns/op           2718524 rows/s             2582 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/16-32           32          34991868 ns/op           2857807 rows/s             2468 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/32-32           32          35449259 ns/op           2820933 rows/s             2572 ts-ids             1000 ts-names
PASS
ok      github.com/VictoriaMetrics/VictoriaMetrics/lib/storage  20.766s
```

And the result after the fix (duplicate metric IDs are not created):

```
user@laptop:~/p/github.com/rtm0/VictoriaMetrics/01/src$ GOMAXPROCS=32 go test -run=^$ -bench=StorageAddHistoricalRowsConcurrently ./lib/storage/ -loggerLevel=ERROR; sync
goos: linux
goarch: amd64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/lib/storage
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkStorageAddHistoricalRowsConcurrently/1-32            24          41935214 ns/op           2384631 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/2-32            28          40869791 ns/op           2446795 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/4-32            30          39618185 ns/op           2524093 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/8-32            26          38496418 ns/op           2597644 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/16-32           32          36164710 ns/op           2765127 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/32-32           32          36914587 ns/op           2708956 rows/s             1000 ts-ids             1000 ts-names
PASS
ok      github.com/VictoriaMetrics/VictoriaMetrics/lib/storage  19.344s
``` 

The benchmark also attemps to measure row insertion rate (see `rows/s` metric values) to see which performance penalty this fix might introduce although the numbers do not look definite enough. The fix introduces a critical section into the insertion path so the it must become slower. Here are the result of a real-world import of a dataset containing `6229440` unique time series.

Import time before fix: `1m56s`
Import time after fix: `2m8s`

That's `10%` slower.

Let's also look at the metrics.

Before the fix, the number of new time series created equals the number the insertion took the slow path:

```
vm_new_timeseries_created_total 15247877
vm_slow_row_inserts_total 15247877
```

After the fix, the number of new time series is exactly `6229440`, but the number of times the slow path was taken is `10%` bigger.

```
vm_new_timeseries_created_total 6229440
vm_slow_row_inserts_total 16775134
```

### Checklist

The following checks are **mandatory**:

- [ x] My change adheres [VictoriaMetrics contributing guidelines](https://docs.victoriametrics.com/contributing/).