@rtm0 2024-06-04
Import and benchmark results after addressing code review remarks:

## Before the fix

Import time: `2m9s`

Metrics:

```
vm_new_timeseries_created_total 15941451
vm_slow_row_inserts_total 15941451
vm_slow_per_day_index_inserts_total 1184
```

Benchmark:

```
user@laptop:~/p/github.com/rtm0/VictoriaMetrics/01/src$ sync; sudo sysctl -w vm.drop_caches=3; GOMAXPROCS=32 go test -run=^$ -bench=StorageAddHistoricalRowsConcurrently ./lib/storage/ -loggerLevel=ERROR
vm.drop_caches = 3
goos: linux
goarch: amd64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/lib/storage
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkStorageAddHistoricalRowsConcurrently/1-32            28          39313259 ns/op           2543671 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/2-32            28          39160446 ns/op           2553597 rows/s             1591 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/4-32            32          39268187 ns/op           2546591 rows/s             1846 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/8-32            31          37131695 ns/op           2693117 rows/s             2522 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/16-32           31          34102306 ns/op           2932353 rows/s             2564 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/32-32           31          38397385 ns/op           2604344 rows/s             2710 ts-ids             1000 ts-names
PASS
ok      github.com/VictoriaMetrics/VictoriaMetrics/lib/storage  22.278s
```

## After the fix

Import time: `2m7s`

Metrics:

```
vm_new_timeseries_created_total 6229440
vm_slow_row_inserts_total 34132319
vm_slow_per_day_index_inserts_total 869
```

Benchmark:

```
user@laptop:~/p/github.com/rtm0/VictoriaMetrics/01/src$ sync; sudo sysctl -w vm.drop_caches=3; GOMAXPROCS=32 go test -run=^$ -bench=StorageAddHistoricalRowsConcurrently ./lib/storage/ -loggerLevel=ERROR
[sudo] password for user: 
vm.drop_caches = 3
goos: linux
goarch: amd64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/lib/storage
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkStorageAddHistoricalRowsConcurrently/1-32            28          40022015 ns/op           2498625 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/2-32            28          39484298 ns/op           2532652 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/4-32            26          39230817 ns/op           2549017 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/8-32            31          37861286 ns/op           2641220 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/16-32           32          33812489 ns/op           2957487 rows/s             1000 ts-ids             1000 ts-names
BenchmarkStorageAddHistoricalRowsConcurrently/32-32           26          39254135 ns/op           2547502 rows/s             1000 ts-ids             1000 ts-names
PASS
ok      github.com/VictoriaMetrics/VictoriaMetrics/lib/storage  19.808s
```