@hagen1778 2024-06-04
Thanks for the patch @rtm0 !

> If we import each day concurrently, then the same new time series will be imported concurrently. 

I know that vmctl does concurrent ingestion of historical data, but in per-metric-name fashion. This helps to properly utilize VM indexes and improves data locality during background merges. Could this be helpful for your case as well?