@rtm0 2024-06-07
Hi @hagen1778!

> I know that vmctl does concurrent ingestion of historical data, but in per-metric-name fashion. This helps to properly utilize VM indexes and improves data locality during background merges. Could this be helpful for your case as well?

That certainly helps. But what I've learned is that duplicate metric IDs is not a big deal. This PR is the result of my getting acquainted with the code base, i.e. mostly educational. However I may increase the accuracy of the numbers on the `Explore Cardinality` page of the vmui. Note that even wit this fix that page would still show inaccurate numbers since another source of inaccuracy is the creation of per-date indexes. This may or may not be addressed in the subsequent PRs.