@hagen1778 2024-06-10
>  However I may increase the accuracy of the numbers on the Explore Cardinality page of the vmui. Note that even wit this fix that page would still show inaccurate numbers since another source of inaccuracy is the creation of per-date indexes. This may or may not be addressed in the subsequent PRs.

True. But I don't think that attempts to make things more accurate for edge cases cost 10% of the ingestion performance for all cases. 

I'm closing this PR as it won't be merged because of the performance penalties it implies. Thanks for your efforts!